Pull protocol FilterConfiguration into separate file
