+++ Sources/FINNFilterImplementation/FilterMarkets/FilterConfiguration.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+import Foundation
+
+protocol FilterConfiguration {
+    func handlesVerticalId(_ vertical: String) -> Bool
+    var preferenceFilterKeys: [FilterKey] { get }
+    var supportedFiltersKeys: [FilterKey] { get }
+    var mapFilterKey: FilterKey? { get }
+    func createRangeFilterFrom(filterData: FilterData) -> RangeFilterInfoType?
+}
